fixing

+++ src/Product_page/review.php
@@ -35,8 +35,6 @@ while($review = $review_result->fetch_assoc()){
       <p><?php echo $review_comment ?></p>
       <div>
         <?php
-        echo $user_id;
-        echo $review_user_id;
         if($user_id == $review_user_id && isset($user_id) || $user_id == 0 && isset($user_id)){
           ?>
           <button class="delete_button" value="Delete" onclick="delete_review('<?php echo $review_id ?>')">Delete</button>
@@ -80,7 +78,7 @@ while($review = $review_result->fetch_assoc()){
         <p><?php echo $comment_comment ?></p>
         <div>
           <?php
-          if($comment_user_id == $user_id || $user_id == 0){
+          if($user_id == $comment_user_id && isset($user_id) || $user_id == 0 && isset($user_id)){
             ?>
             <button class="delete_button" value="Delete" onclick="delete_comment('<?php echo $comment_id ?>')">Delete</button>
             <?php
